telefon-takip-app/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
gspread==5.12.0
oauth2client==4.1.3
Pillow==10.0.0
schedule==1.2.0
openai==1.3.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
Flask==2.3.3
APScheduler==3.10.4 